Change Detection Loop Bug
Fixed a bug where infinite change detection would occur on the `MapComponent`. This stemmed from the `Site.getMapMarker()` function which returns a new object each time it is called. If the function is called directly into a components HTML, change detection will continue to evaluate it as a new object and infinitely load eventually lagging out the page. To solve, I've created a `sanitizeMapMarkers()` function which can be used to fix change detection for single, and multiple marker use cases.