Fix unit tests (#362)
* Cleaned up broken unit tests

Had an edge case where it appeared that I could not mock the MapComponent. This turns out to stem from the fact that unit tests cannot override the components for modules which are not directly tied to the component. In this case, the custom formly inputs are loaded as part of a smaller module, which means I cannot override the MapComponent

* Mock services (#363)

* Extract service/guard to modules

* Removed testAppInitializer

* Removed testBawServices

* Mocked services with SpyObjects

* Created ApiErrorDetails generator