Re-implement PageDecorator
Angular uses "decorators" which are actually converted by their compiler into something else which allows them to maintain AOT. If it detects any other decorators on an angular injectable, the custom decorator is removed. Otherwise any other custom decorators are kept as long as they do not have @angular/core in its path.

Because of this, the PageDecorator is compiled out in the production build of workbench client. This issue forced a rewrite of how PageComponents are implemented.

Model decorators are untouched as they do not involve angular injection directives.