Annotation model (#567)
* Cleanup route

Forgot to change back the analysis job view url

* Attempt to filter recordings by audio event creator id

* Initial look at annotations table

* Commit code across computers

* Cleanup AssociationDecorator types

Cosmetic changes to ServiceToken to attempt to clean them up

* Working my/their-annotations page

Modified the many UnresolvedModels array to contain one UnresolvedModel. This maintains the output as an array, whilst also giving the ability to determine if the value is still loading.
Would like to make modifications to the HasMany decorator so that it short circuits requests which contain no ids

* Reduce HasMany excess queries

HasMany was making requests to the API even when it had an empty array of ids. This has now been prevented by making it terminate early and return the failureValue

* Updated table decriptions

Updated the description above paged table template using components so that they will display the number of models in the current page, and the total number of models.
Discoved admin tag groups list page is broken. Api endpoint is wrong

* Created isUnresolved pipe

* Created helper pipes

* Update src/app/models/AssociationDecorators.ts

Co-authored-by: Anthony Truskinger <anthony.truskinger@gmail.com>

* Making tests compilable

* Implemented pipe tests

* Updated AssociationDecorator tests

Breaking change to how association decorators work. They no longer accept multiple responses because it has no use case, and complicates an already complicated function. This logic can be easily re-implemented. Tests for this scenario have been removed.

* Cleanup remaining tests

* Implemented dummy links

* Re-implement loading spinner

* Remove spinner pipe

Co-authored-by: Anthony Truskinger <anthony.truskinger@gmail.com>