Fix visual bug when menued page collapses
Visual bug existed which would make the action menu have a large unused space between it and the secondary menu. This only applied to the collapsed menu layout, and could take up a large segment of the page. Issue stemmed from using the bootstrap flex system, which would make the action menu use the full flex height, instead of changing the height to match the content.