Fix 3 second delay on ssr
Rxjs `delay` operator on the loading bar for http requests caused zonejs to wait for the first trigger of the observable before returning the build. This was fixed by changing the delay to use the global config for ngx-loading. ngx-loading disables its use in SSR to prevent this specific issue.