Fixed print styles on Firefox
Turns out that window.print print styles on firefox were broken because
it was printing in "fit to page" mode instead of "scale 100%" (which is
what Ctrl+P was printing in). I was somehow able to trick my user agent
into thinking I was in scale 100% mode with window.print, so it showed I
was in scale 100%, but it was acutally "fit to page".

Seems pretty straight forward and I might be able to file my 4th bug
report with Mozilla if I can reproduce how I tricked my user agent.