Component resize now triggers resize event
Vega lite only triggers it's internal resize event when window.resize is
called. By triggering the vega lite resize event when the component
resizes, it resizes if it's container shape changes.

This fixes the charts being permenantly hidden if the chart is hidden on
page load, fixes text scaling of charts under certain conditions (on
certain browsers & platforms) and fixes a bug where padding whitespace would be
present on printing if the container size was not equal to the chart svg
size.