Quick hacky fix for project scoping
The correct fix here will be to use route resolvers and add the project,
region, and site filters to the annotation search parameters data model.

This is because I cannot directly use the association decorators because
they are async, and do not await. Therefore returning before the full
model can be fetched.