Partial hacky fixes for verification component
This commit fixes the tests for the api backed verification component.
However, it does not assert what createOrUpdate is called with (hence
the hacky comment)

A following commit will be added to assert that the createOrUpdate
service method is called with the correct decision content and clean up
a lot of the hacky code.

This commit is being pushed to test in CI