Remove protractor
When this project originally started, we wanted to use protractor for
e2e tests.

However, we never wrote one e2e test, never enabled in CI, and never
worked.

This commit removes protractor as a dependency which makes test
migration scripts easier to run, removes a dependency, and removes a lot
of deprecation warnings.