Move resolver-common into sub-directory
Instead of cluttering the root of the baw-api directory, there is now a
"resolvers" directory so that I can start refactoring.

I have made this an independent commit so that git can easily recognise
that the file was moved and can more efficantly produce diffs and clone
the repo faster.