Remove missing tag warning
No one actually asked for this. I thought that it'd be simple and useful
to add, but it turned out that it's quite hard to make a good solution
to extract warnings.

This is because warnings are generated on a per-event basis, meaning
that we'd have to process the entire uploaded file client side (not
very performant).