Change order of decision buttons for hacky fix
Two skip buttons were being created.

I just remove the second skip button and re-ordered the buttons. I don't
like the semantics of it because the buttons now read right to left, but
it's better than a broken UI or not shipping the feature in time.

I suspect this is an upstream bug to do with us only checking if a skip
button exists or not in the firstUpdated hook instead of checking every
update, meaning that SPA's can init without a skip button, causing a
skip button to be added, and then an explicit skip button is added
and we don't re-check.
This problem didn't arise during SSR rendering/fresh reloads which is
why I didn't see it during dev.