Do not send authentication with auth requests
This also refactors out the http context WITH_AUTHENTICATION and
consolidates authentication logic into the with_credentials option so
that there is one source of truth.

Note that this doesn't work although the requests look correct.
I'm going to come back later when I'm not looking at the same code.