Custom karma middleware to serve assets with COEP
The Karma "customHeaders" config option only gets applied to the testing
files and not files served through the "files" config option.

We use a custom middleware to apply all of the config.customHeaders to
the config.files

Note that we use config.customHeaders as the source of truth for the
file headers so that we don't have multiple points of maintainance.
Updating config.customHeaders should transparently apply to all assets.