Use application builder for tests
I did this because the webpack middleware was throwing errors.

Additionally, using webpack for tests while the real build uses esbuild
introduces bug risks.

I think Karma or esbuild might be doing something weird to the audio
files, because sometimes the test runner does a full reload (as if the
file changed), and the newly served file has different bytes from the
source file and fails to play in VLC or the spectrogram.
This isn't ideal, but at least the tests now run.