fix url directive tests
Webpack and esbuild URL constructors were behaving slightly different.
Webpack builds were returning document.hostURI as the test runners
document location, while esbuild was referring to the iframes uri.

This caused some problems with the expected urls in tests.

Note this was only a bug in the test fixture, not in the url directive.
The url directive has remained unchanged.