Transition to "load" state if map takes > 300ms
For fast loading maps, we don't enter a loading state because it'd flash
white, giving a jarring experience.

For slow requests, we still want to provide feedback to the user.

Note that we never used to have a "data fetching" state. This is a new
feature that I added because waiting 2 seconds for the a2o map to load
after typing in a site name looked really slow.
Note that if you type in a site name that narrows to a small subset of
sites, we never transition into this fetching state.