Revert marker cache
I thought that I could cache marker content between groups, so that each
marker grouping could share an element reference for content (so that we
wouldn't re-create heaps of duplicate pin elements).

However, this didn't seem to work, and only one pin in the group would
have the correct styling.

I still like the idea, so I added a TODO.

However, in the spirit of merging this PR, I'm happy to merge without
pin content caching, because it wasn't really slow to begin with.