Faster detectChanges helper
We have a custom detectChanges helper that waits for Angular & Lit to
complete updates.

However, it was extremely slow, causing a lot of tests to become flakey
because of timeouts.

By improving the performance of this helper, we can fix these flakey
tests and also the speed of our test suite.