Ensure unique ids in mock data (#2496)
Some tests were flaky because there was a ~1/100 probability to generate mock data with the same id.

I have observed this causing issues when models attempt to get the ids of sub-models and map the sub-model ids into a Set() which causes some of the expected ids disappear.

Note that this doesn't fix all the flaky tests, it just fixes one potential flake point.