Attempt to use library build
I assumed that we would want a library build because then we can get all
of the Angular native library helpers such as automatically creating a
package.json, creating typings, etc...

However, looking at the build output, it ended up being too
Angular-specific and required Angular as a peer dependency.

This isn't ideal for our use-case because we just want to have a
singular file that we can add.

Comparing the two outputs, I think we want an application build with a
createApplication (not bootstrapApplication) entrypoint.

[skip ci]