Fix CI
I'm 99% sure it was a problem with the bawImage decorator

My **theory** is that:

In FF145 it looks like the logic for optional chaining changed.

Essentially, undefined?.get("xyz") used to short circut at the
"undefined" meaning that the "get" never attempted to call.

However, in new FF versions, it would not short circut, and it would try
to call "undefined" as a function.
By using the ?.("xyz") invokation of model?.injector, it should fix this
bug.