Fix verification parameter model parsing url tags
Both the verification parameters and the annotation search parameters
parsed the "tags" url parameter.
This is because the verification parameters need to know what tags were
filtered for so that it can do its tag priority correctly.

However, this was flawed because it meant that there were two sources of
truth as to what the filtered "tags" were.
Meaning that sometimes if you updated the "tags" it would update the
annotation search parameter model without updating the verification
model.
This caused some inconsistent state from having two sources of truth.

E.g. The url parameters would prioritise the verification parameter
"tags" over the search parameters, but the filters would use the
annotation search filters.

---

The fix I applied in this commit was using the
annotationSearchParameters as the source of truth for what tags were
filtered and using it as a parameter to the verificationParameters
tagPriority method.

However, this fix is obviously not ideal, so I'm thinking that I'll make
this tagPriority function a method on the verification.service so that I
can treat these two parameter models as independent, and not needing to
know about the other.

Once I turn tagPriority into a service method, the consumer (the
verification page) can handle tag priority filter construction.

Fixes: #2512

[skip ci]