Use verification component method for tag priority
This moves tag priority out of the verificationParameters where it
didn't belong because I want the verificationParameters to not need any
knowledge of the search parameters.

Unfortunately, I still do need the search parameters for the
verification filter form because I want to show the filtered tags as a
placeholder default value in the "tags to verify" input.