Fix flaky add-annotations tests
I found a flaky tests that was caused by not correctly awaiting
association resolution.

Associations are resolved in async, but because these tests were not run
in a fakeAsync zone, there was no guarantee that the promises would
resolve before the assertion, meaning that sometimes the associations
would (rarely) not resolve, meaning that the "annotation preview" table
was partially incomplete, causing the test to fail.

Related to: #2495