Fix flaky add-annotations tests (#2531)
Associations are resolved in async, but because these tests were not run in a fakeAsync zone, there was no guarantee that the promises would resolve before the assertion, meaning that sometimes the associations would (rarely) not resolve, meaning that the "annotation preview" table was partially incomplete, causing the test to fail.

Additionally, ImportedAudioEvent.tags had a 1/10 chance to be an empty array in a test
that expected populated tag models.

Related to: #2495