Fix flaky verification tests
The annotation model could sometimes create with zero tags, meaning that
the verification task was marked as "not required", meaning that the
decision would hang and timeout.

While trying to find the flaky test, I also mocked the verification
grids `getPage` callback during testing so that it now only returns one
page. This stops the gridPageFetcher from running in the background
pre-fetching many pages in advance, when most tests just need the first
page of items to assert.

This significantly slowed down the tests, and could cause the test suite
to fail if the verification tests were run at the end of the test suite
and karma closed without correctly closing the gridPageFetcher stream
(I've created an upstream issue in web-components to fix this issue).