Attempt subset removal and realise it's very hard
I was hoping that if you removed an audio event import all of its files
would be automatically removed.

However, this is quite hard without async associations because we only
have the id of the import file, meaning that to resolve the full import
file model and to get the `audioEventImportId` is possible although
quite difficult for the benefit it would bring.

I'll hold out on this feature until we have async associations.