Working map implementation
Skipping CI because I was able to simplify the filters so instead of
each file having its own or: [ { eventImportFileId: { eq: x } } ]
condition, it now uses an `in` condition.

This was actually easier because I already have all of the files in a
set, so I just needed to convert to an array.

[skip ci]